# src/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: AsapJoin Carpool Platform API
  description: |
    API REST complète pour la plateforme de covoiturage AsapJoin.
    Gère les trajets, réservations, livraisons de colis, paiements Stripe,
    wallet interne, politiques d'annulation, payouts et messagerie.
  version: 1.0.0
  contact:
    name: AsapJoin Team
servers:
  - url: http://localhost:3000
    description: Development
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input
        traceId:
          type: string
          format: uuid
        details:
          type: array
          items:
            type: object
...
    CancelPreview:
      type: object
      properties:
        allowed:
          type: boolean
          description: Whether cancellation is allowed
        policy_id:
          type: string
          nullable: true
        policy_name:
          type: string
          nullable: true
        refund_amount_cents:
          type: integer
          description: Amount to refund to customer (in cents)
        cancellation_fee_cents:
          type: integer
          description: Cancellation fee retained (in cents)
        driver_reversal_cents:
          type: integer
          description: Amount reversed from driver wallet (in cents)
        driver_compensation_cents:
          type: integer
          description: Compensation to driver for late cancellation (in cents)
        original_amount_cents:
          type: integer
          description: Original gross payment amount (in cents)
        message:
          type: string
          description: Human-readable message
        reason_code:
          type: string
          nullable: true
          description: Machine-readable reason code
          enum: [CANCELLATION_NOT_ALLOWED, INVALID_REFUND_POLICY, REFUND_REQUEST_WINDOW_EXPIRED]
    CancelResult:
      type: object
      properties:
        cancellation_request_id:
          type: string
        status:
          type: string
          enum: [approved, refunded]
        refund_amount_cents:
          type: integer
        cancellation_fee_cents:
          type: integer
        stripe_refund_id:
          type: string
          nullable: true
        message:
          type: string
    RefundPolicy:
      type: object
      properties:
        id:
          type: string
        resource_type:
          type: string
          enum: [booking, delivery]
        actor_role:
          type: string
          enum: [passenger, sender, driver, admin]
        name:
          type: string
        active:
          type: boolean
        min_hours_before_departure:
          type: integer
          description: Minimum hours before departure to allow cancellation
        refund_request_deadline_hours:
          type: integer
          description: Max hours after event to request refund (0 = no deadline)
        cancellation_fee_fixed_cents:
          type: integer
        cancellation_fee_percent:
          type: number
          format: decimal
        refund_percent_to_customer:
          type: number
          format: decimal
        driver_compensation_percent:
          type: number
          format: decimal
        applies_when_statuses:
          type: string
          description: CSV of statuses where this policy applies
        priority:
          type: integer
        notes:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    RefundPolicyInput:
      type: object
      required: [resource_type, actor_role, name, applies_when_statuses]
      properties:
        resource_type:
          type: string
          enum: [booking, delivery]
        actor_role:
          type: string
          enum: [passenger, sender, driver, admin]
        name:
          type: string
        min_hours_before_departure:
          type: integer
          default: 0
        refund_request_deadline_hours:
          type: integer
          default: 0
        cancellation_fee_fixed_cents:
          type: integer
          default: 0
        cancellation_fee_percent:
          type: number
          default: 0
        refund_percent_to_customer:
          type: number
          default: 100
        driver_compensation_percent:
          type: number
          default: 0
        applies_when_statuses:
          type: string
          description: CSV of statuses (e.g. "pending,accepted,paid")
        priority:
          type: integer
          default: 0
        notes:
          type: string
    CancellationRequest:
      type: object
      properties:
        id:
          type: string
        resource_type:
          type: string
          enum: [booking, delivery]
        resource_id:
          type: string
        actor_user_id:
          type: string
        actor_role:
          type: string
          enum: [passenger, sender, driver, admin]
        reason:
          type: string
          nullable: true
        original_amount_cents:
          type: integer
        calculated_refund_cents:
          type: integer
        calculated_fee_cents:
          type: integer
        driver_reversal_cents:
          type: integer
        commission_reversal_cents:
          type: integer
        driver_compensation_cents:
          type: integer
        policy_id:
          type: string
          nullable: true
        status:
          type: string
          enum: [requested, approved, rejected, refunded, closed]
        stripe_refund_id:
          type: string
          nullable: true
        refund_id:
          type: string
          nullable: true
        is_admin_override:
          type: boolean
        created_at:
          type: string
          format: date-time
        processed_at:
          type: string
          format: date-time
          nullable: true
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
paths:
...
  /api/v1/bookings/{id}/cancel-preview:
    get:
      tags: [Cancellations]
      summary: Preview booking cancellation (dry-run)
      description: |
        Returns the refund calculation for a potential cancellation.
        Uses the paramétrable refund policy engine to determine:
        - Whether cancellation is allowed
        - Refund amount
        - Cancellation fees
        - Driver compensation
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Cancellation preview
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CancelPreview'
        '404':
          description: Booking not found
  /api/v1/bookings/{id}/cancel:
    post:
      tags: [Cancellations]
      summary: Cancel a booking with refund workflow
      description: |
        Executes the full cancellation workflow:
        1. Validates actor role and permissions
        2. Resolves applicable refund policy
        3. Validates cancellation and refund request windows
        4. Computes refund amounts
        5. Executes Stripe refund if applicable
        6. Writes append-only ledger corrections
        7. Updates booking status
        8. Creates cancellation_request record
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  description: Reason for cancellation
      responses:
        '200':
          description: Cancellation executed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CancelResult'
        '400':
          description: Business rule violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                already_cancelled:
                  value:
                    code: ALREADY_CANCELLED
                    message: This item has already been cancelled
                not_allowed:
                  value:
                    code: CANCELLATION_NOT_ALLOWED
                    message: Cancellation is not allowed in current status
                window_expired:
                  value:
                    code: REFUND_REQUEST_WINDOW_EXPIRED
                    message: The refund request deadline has passed
                invalid_policy:
                  value:
                    code: INVALID_REFUND_POLICY
                    message: No applicable refund policy found
...
  /api/v1/admin/refund-policies:
    get:
      tags: [Refund Policies]
      summary: List all refund policies
      description: Admin/support can list all paramétrable refund policies.
      parameters:
        - name: resource_type
          in: query
          schema:
            type: string
            enum: [booking, delivery]
        - name: actor_role
          in: query
          schema:
            type: string
            enum: [passenger, sender, driver, admin]
        - name: active
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: List of refund policies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RefundPolicy'
    post:
      tags: [Refund Policies]
      summary: Create a new refund policy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefundPolicyInput'
      responses:
        '201':
          description: Policy created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefundPolicy'
  /api/v1/admin/refund-policies/{id}:
    patch:
      tags: [Refund Policies]
      summary: Update a refund policy
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefundPolicyInput'
      responses:
        '200':
          description: Policy updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefundPolicy'
  /api/v1/admin/refund-policies/{id}/activate:
    post:
      tags: [Refund Policies]
      summary: Activate a refund policy
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Policy activated
  /api/v1/admin/refund-policies/{id}/deactivate:
    post:
      tags: [Refund Policies]
      summary: Deactivate a refund policy
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Policy deactivated

  # ─── Admin Refund Override ───
  /api/v1/admin/refunds/override:
    post:
      tags: [Refund Policies]
      summary: Admin override refund
      description: |
        Admin can force a refund on any booking or delivery, bypassing the standard policy.
        Creates Stripe refund, ledger corrections, and cancellation_request record.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [resource_type, resource_id, refund_amount_cents]
              properties:
                resource_type:
                  type: string
                  enum: [booking, delivery]
                resource_id:
                  type: string
                refund_amount_cents:
                  type: integer
                  description: Amount to refund in cents
                reason:
                  type: string
                override_policy:
                  type: boolean
                  default: true
      responses:
        '200':
          description: Override refund executed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CancelResult'
        '400':
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                no_payment:
                  value:
                    code: NO_PAYMENT_FOUND
                    message: No succeeded payment found for this resource
                already_refunded:
                  value:
                    code: REFUND_ALREADY_PROCESSED
                    message: This resource has already been fully refunded

  # ─── Admin Cancellation Requests ───
  /api/v1/admin/cancellation-requests:
    get:
      tags: [Refund Policies]
      summary: List cancellation requests
      description: Admin/support can view all cancellation requests with their status and financial details.
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [requested, approved, rejected, refunded, closed]
        - name: resource_type
          in: query
          schema:
            type: string
            enum: [booking, delivery]
      responses:
        '200':
          description: List of cancellation requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CancellationRequest'